+++ src/adminfilters/__init__.py
@@ -1,4 +1,4 @@
 default_app_config = 'adminfilters.apps.Config'
 NAME = 'django-adminfilters'
-VERSION = __version__ = '2.0.1'
+VERSION = __version__ = '2.0.3'
 __author__ = 'sax'