+++ src/adminfilters/utils.py
@@ -55,6 +55,8 @@ def get_field_by_name(model, name):
 
 def get_all_field_names(model):
     from itertools import chain
+    if not model:
+        raise ValueError("'model' must be a Model instance")
     return list(set(chain.from_iterable((field.name, field.attname)
                                         if hasattr(field, 'attname') else (field.name,)
                                         for field in model._meta.get_fields()
@@ -81,11 +83,21 @@ def get_field_by_path(model, field_path):
     """
     parts = field_path.split('.')
     target = parts[0]
+    if not model:
+        raise ValueError("'model' must be a Model instance")
+
     if target in get_all_field_names(model):
         field_object, model, direct, m2m = get_field_by_name(model, target)
         if isinstance(field_object, models.ForeignKey):
             if parts[1:]:
-                return get_field_by_path(field_object.related_model, '.'.join(parts[1:]))
+                while True:
+                    if field_object.related_model is None:
+                        break
+                    fk = get_field_by_path(field_object.related_model, '.'.join(parts[1:]))
+                    if fk is None:
+                        break
+                    field_object = fk
+                return field_object
             else:
                 return field_object
         else:
@@ -108,11 +120,11 @@ def get_field_type(model, field_path):
     return field, lookup, field_type
 
 
-def cast_value(v, fld, multiple):
-    if isinstance(fld, (BooleanField,)):
+def cast_value(v, fld, lookup):
+    if isinstance(fld, (BooleanField,)) or lookup in ['isnull']:
         func = parse_bool
     else:
         func = fld.to_python
-    if multiple:
+    if lookup in ['in']:
         return [func(e) for e in v.split(',')]
     return func(v)